Mark Lex and Yacc prerequisites order-only
This makes is so that the C intermediates will only be remade if the
Lex or Yacc source is completely missing. To force remaking those, run
`make realclean` beforehand.

This should help remove the dependency on Lex and Yacc, allowing the
vendored generated intermediate files to be used by the build process.